work on umod functionality a bit
