--- UmodServer/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="DotNetZip" version="1.16.0" targetFramework="net48" />
-  <package id="System.IO.Compression" version="4.3.0" targetFramework="net461" />
-</packages>
